scripts/csv2sql.sh: bizarre id of 1 for attrs context_class_id bug
for some reason, it was assumed that projects would only have one
context class and that its id would be 1.

Now we have an explicit column for the context class ID, and we
use that datum to render the SQL.

db/ods/template.ots: updated template with the newly required columns
in cat_attribute table.