site/style.css: Inkscape compatibility and explicit font selection
Font had to be set to an explicit font and not a semantic "sans-serif"
as each SVG interpreter may take a different font for sans-serif, and
it was messing up the text boxes.

Inkscape has trouble with gradient stops that use alpha in the color,
but a simple move to stop-opacity fixed the problem.